+++ watch/tasks/fusion/datamodules/temporal_sampling/utils.py
@@ -1,4 +1,5 @@
 import numpy as np
+import ubelt as ub
 
 
 def guess_missing_unixtimes(unixtimes):
@@ -80,6 +81,7 @@ def coerce_time_kernel(pattern):
         >>>     '-60s,0s,60s',
         >>>     '-1d,-60s,20s,60s,1d',
         >>>     '1,1,1,1,1',
+        >>>     '(1,1,1,1,1)',
         >>> ]
         >>> for pattern in valid_patterns:
         >>>     kernel = coerce_time_kernel(pattern)
@@ -88,6 +90,7 @@ def coerce_time_kernel(pattern):
         kernel = [-31536000.0, -2592000.0, -86400.0, 0.0, 86400.0, 2592000.0, 31536000.0]
         kernel = [-60.0, 0.0, 60.0]
         kernel = [-86400.0, -60.0, 20.0, 60.0, 86400.0]
+        kernel = [1.0, 1.0, 1.0, 1.0, 1.0]
         kernel = [1.0, 1.0, 1.0, 1.0, 1.0]
         >>> import pytest
         >>> invalid_patterns = [
@@ -100,14 +103,19 @@ def coerce_time_kernel(pattern):
         >>> with pytest.raises(TypeError):
         >>>     coerce_time_kernel(3.14)
     """
+    from watch.tasks.fusion.datamodules.temporal_sampling.time_kernel_grammar import parse_multi_time_kernel
     from watch.utils.util_time import coerce_timedelta
-    import numpy as np
-    import ubelt as ub
     if isinstance(pattern, str):
-        kernel_deltas = pattern.split(':')[-1].split(',')
+        if '(' in pattern:
+            multi_kernel = parse_multi_time_kernel(pattern)
+            assert len(multi_kernel) == 1, 'only expecting a single kernel here'
+            kernel_deltas = multi_kernel[0]
+        else:
+            kernel_deltas = pattern.split(',')
     elif ub.iterable(pattern):
         kernel_deltas = pattern
     else:
+        print(f'error: pattern={pattern}')
         raise TypeError(type(pattern))
     parsed = [coerce_timedelta(d) for d in kernel_deltas]
     kernel = np.array([v.total_seconds() for v in parsed])
@@ -115,3 +123,60 @@ def coerce_time_kernel(pattern):
     if not np.all(diffs >= 0):
         raise ValueError('Inputs must be in ascending order')
     return kernel
+
+
+def coerce_multi_time_kernel(pattern):
+    """
+    Obtain a list of time kernels from user input.
+
+    Example:
+        >>> from watch.tasks.fusion.datamodules.temporal_sampling.utils import *  # NOQA
+        >>> import ubelt as ub
+        >>> valid_patterns = [
+        >>>     '(-1d,0,1d),(1,2)',
+        >>>     '1,1,1',
+        >>>     '(1,1,1,1,1)',
+        >>>     ['-1,0,+1', '0'],
+        >>> ]
+        >>> for pattern in valid_patterns:
+        >>>     multi_kernel = coerce_multi_time_kernel(pattern)
+        >>>     recon = coerce_multi_time_kernel(multi_kernel)
+        >>>     a = [r.tolist() for r in recon]
+        >>>     b = [r.tolist() for r in multi_kernel]
+        >>>     assert a == b, 'should be idempotent'
+        >>>     print('multi_kernel = {}'.format(ub.urepr(multi_kernel, nl=1)))
+        multi_kernel = [
+            np.array([-86400.,      0.,  86400.], dtype=np.float64),
+            np.array([1., 2.], dtype=np.float64),
+        ]
+        multi_kernel = [
+            np.array([1., 1., 1.], dtype=np.float64),
+        ]
+        multi_kernel = [
+            np.array([1., 1., 1., 1., 1.], dtype=np.float64),
+        ]
+        multi_kernel = [
+            np.array([-1.,  0.,  1.], dtype=np.float64),
+            np.array([0.], dtype=np.float64),
+        ]
+    """
+    if pattern is None:
+        return [None]
+    from watch.tasks.fusion.datamodules.temporal_sampling.time_kernel_grammar import parse_multi_time_kernel
+    if isinstance(pattern, str):
+        multi_kernel = parse_multi_time_kernel(pattern)
+    elif ub.iterable(pattern):
+        if len(pattern) == 0:
+            multi_kernel = []
+        else:
+            first = pattern[0]
+            if ub.iterable(first):
+                # Assume we are given a list of pre-parsed kernels
+                multi_kernel = pattern
+            else:
+                # Assume we are given a list of parseable kernels
+                multi_kernel = [coerce_time_kernel(p) for p in pattern]
+    else:
+        print(f'error: pattern={pattern}')
+        raise TypeError(type(pattern))
+    return multi_kernel